+++ controllers/mirrorpeer/mirrorpeer_controller.go
@@ -18,7 +18,6 @@ package mirrorpeer
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"log/slog"
 	"slices"
@@ -31,7 +30,6 @@ import (
 	"github.com/red-hat-storage/odf-multicluster-orchestrator/version"
 
 	ramenv1alpha1 "github.com/ramendr/ramen/api/v1alpha1"
-	ocsv1 "github.com/red-hat-storage/ocs-operator/api/v4/v1"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -42,7 +40,6 @@ import (
 	"k8s.io/klog/v2"
 	"k8s.io/utils/ptr"
 	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
-	workv1 "open-cluster-management.io/api/work/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -63,12 +60,11 @@ type MirrorPeerReconciler struct {
 }
 
 const (
-	mirrorPeerFinalizer                 = "hub.multicluster.odf.openshift.io"
-	AddonVersionAnnotationKey           = "multicluster.odf.openshift.io/version"
-	StorageClusterPeerNameAnnotationKey = "ocs.openshift.io/storage-cluster-peer"
-	MirrorPeer                          = "MirrorPeer"
-	ManagedClusterAddOn                 = "ManagedClusterAddOn"
-	AddonApiVersion                     = "addon.open-cluster-management.io/v1alpha1"
+	mirrorPeerFinalizer       = "hub.multicluster.odf.openshift.io"
+	AddonVersionAnnotationKey = "multicluster.odf.openshift.io/version"
+	MirrorPeer                = "MirrorPeer"
+	ManagedClusterAddOn       = "ManagedClusterAddOn"
+	AddonApiVersion           = "addon.open-cluster-management.io/v1alpha1"
 )
 
 // +kubebuilder:rbac:groups=multicluster.odf.openshift.io,resources=mirrorpeers,verbs=get;list;watch;update;patch;delete
@@ -289,19 +285,19 @@ func (r *MirrorPeerReconciler) ensureReplicationReady(ctx context.Context, logge
 		return ctrl.Result{}, nil
 	}
 
-	if err := createStorageClusterPeer(ctx, r.Client, logger, mirrorPeer, clientInfoMap); err != nil {
+	if err := odf.CreateStorageClusterPeer(ctx, r.Client, logger, mirrorPeer, clientInfoMap); err != nil {
 		logger.Error("Failed to create StorageClusterPeer", "error", err)
 		mirrorPeer.Status.Message = multiclusterv1alpha1.ConfigurationFailed
 		return ctrl.Result{}, err
 	}
 
-	if err := createManifestWorkForClusterPairingConfigMap(ctx, r.Client, logger, mirrorPeer, clientInfoMap); err != nil {
+	if err := odf.CreateManifestWorkForClusterPairingConfigMap(ctx, r.Client, logger, mirrorPeer, clientInfoMap); err != nil {
 		logger.Error("Failed to create ManifestWork for ClusterPairingConfigMap", "error", err)
 		mirrorPeer.Status.Message = multiclusterv1alpha1.ConfigurationFailed
 		return ctrl.Result{}, err
 	}
 
-	providerModePeeringDone, err := isProviderModePeeringDone(ctx, r.Client, r.Logger, mirrorPeer, clientInfoMap)
+	providerModePeeringDone, err := odf.IsProviderModePeeringDone(ctx, r.Client, r.Logger, mirrorPeer, clientInfoMap)
 	if err != nil {
 		mirrorPeer.Status.Message = multiclusterv1alpha1.PeeringIncomplete
 		return ctrl.Result{}, fmt.Errorf("failed to check if provider mode peering is correctly done %w", err)
@@ -350,7 +346,7 @@ func (r *MirrorPeerReconciler) resolveS3Profile(ctx context.Context, logger *slo
 
 	if hasStorageClientRef {
 		var err error
-		secretName, namespace, err = GetNamespacedNameForClientS3Secret(peerRef, mirrorPeer, clientInfoMap)
+		secretName, namespace, err = odf.GetNamespacedNameForClientS3Secret(peerRef, mirrorPeer, clientInfoMap)
 		if err != nil {
 			return nil, "", fmt.Errorf("failed to get namespace for s3 secret: %w", err)
 		}
@@ -387,239 +383,6 @@ func (r *MirrorPeerReconciler) resolveS3Profile(ctx context.Context, logger *slo
 	}
 
 	return profile, secretName, nil
-}
-
-func createManifestWorkForClusterPairingConfigMap(ctx context.Context, client client.Client, logger *slog.Logger, mirrorPeer *multiclusterv1alpha1.MirrorPeer, clientInfoMap map[string]string) error {
-	logger.Info("Starting to create ManifestWork for cluster pairing ConfigMap")
-
-	logger.Info("Fetched client info ConfigMap successfully")
-	items := mirrorPeer.Spec.Items
-
-	ci1, err := odf.GetClientInfoFromConfigMap(clientInfoMap, utils.GetKey(items[0].ClusterName, items[0].StorageClusterRef.Name))
-	if err != nil {
-		logger.Error("Failed to get client info from ConfigMap for the first cluster")
-		return err
-	}
-
-	logger.Info("Fetched client info for the first cluster", "ClientInfo", ci1)
-
-	ci2, err := odf.GetClientInfoFromConfigMap(clientInfoMap, utils.GetKey(items[1].ClusterName, items[1].StorageClusterRef.Name))
-	if err != nil {
-		logger.Error("Failed to get client info from ConfigMap for the second cluster")
-		return err
-	}
-
-	logger.Info("Fetched client info for the second cluster", "ClientInfo", ci2)
-	logger.Info("Updating provider ConfigMap with client pairing", "ProviderClient1", ci1.ClientID, "PairedClient1", ci2.ClientID)
-	if err := updateProviderConfigMap(logger, ctx, client, mirrorPeer, ci1, ci2); err != nil {
-		return err
-	}
-
-	logger.Info("Updating provider ConfigMap with client pairing", "ProviderClient2", ci2.ClientID, "PairedClient2", ci1.ClientID)
-	if err := updateProviderConfigMap(logger, ctx, client, mirrorPeer, ci2, ci1); err != nil {
-		return err
-	}
-
-	logger.Info("Successfully created ManifestWork for cluster pairing ConfigMap")
-	return nil
-}
-
-// updateProviderConfigMap updates the ConfigMap on the provider with the new client pairing
-func updateProviderConfigMap(logger *slog.Logger, ctx context.Context, client client.Client, mirrorPeer *multiclusterv1alpha1.MirrorPeer, providerClientInfo odf.ClientInfo, pairedClientInfo odf.ClientInfo) error {
-	providerName := providerClientInfo.ProviderInfo.ProviderManagedClusterName
-	manifestWorkName := "storage-client-mapping"
-	manifestWorkNamespace := providerName
-
-	logger.Info("Fetching existing ManifestWork for provider", "Namespace", manifestWorkNamespace)
-	manifestWork, err := utils.GetManifestWork(ctx, client, manifestWorkName, manifestWorkNamespace)
-	var configMap *corev1.ConfigMap
-
-	if err != nil {
-		if k8serrors.IsNotFound(err) {
-			logger.Info("ManifestWork not found; creating a new ConfigMap")
-			configMap = &corev1.ConfigMap{
-				TypeMeta: metav1.TypeMeta{
-					Kind:       "ConfigMap",
-					APIVersion: corev1.SchemeGroupVersion.String(),
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "storage-client-mapping",
-					Namespace: providerClientInfo.ProviderInfo.NamespacedName.Namespace,
-					Annotations: map[string]string{
-						StorageClusterPeerNameAnnotationKey: getStorageClusterPeerName(pairedClientInfo.ProviderInfo.ProviderManagedClusterName),
-					},
-				},
-				Data: make(map[string]string),
-			}
-		} else {
-			return fmt.Errorf("failed to get ManifestWork: %w", err)
-		}
-	} else {
-		logger.Info("Found existing ManifestWork, decoding ConfigMap")
-		if len(manifestWork.Spec.Workload.Manifests) == 0 {
-			return fmt.Errorf("ManifestWork %s has no manifests", manifestWorkName)
-		}
-		objJson := manifestWork.Spec.Workload.Manifests[0].RawExtension.Raw
-		configMap, err = utils.DecodeConfigMap(objJson)
-		if err != nil {
-			return fmt.Errorf("failed to decode ConfigMap: %w", err)
-		}
-	}
-
-	logger.Info("Updating ConfigMap with paired client info", "ProviderClientID", providerClientInfo.ClientID, "PairedClientID", pairedClientInfo.ClientID)
-	configMap.Data[providerClientInfo.ClientID] = pairedClientInfo.ClientID
-
-	updatedObjJson, err := json.Marshal(configMap)
-	if err != nil {
-		return fmt.Errorf("failed to marshal updated ConfigMap: %w", err)
-	}
-
-	ownerRef := metav1.OwnerReference{
-		APIVersion: mirrorPeer.APIVersion,
-		Kind:       mirrorPeer.Kind,
-		Name:       mirrorPeer.Name,
-		UID:        mirrorPeer.UID,
-	}
-
-	logger.Info("Creating or updating ManifestWork with updated ConfigMap")
-	_, err = utils.CreateOrUpdateManifestWork(ctx, client, manifestWorkName, manifestWorkNamespace, updatedObjJson, []workv1.ManifestConfigOption{}, ownerRef)
-	if err != nil {
-		return fmt.Errorf("failed to update ManifestWork for provider %s: %w", providerName, err)
-	}
-
-	logger.Info("Successfully updated ManifestWork for provider", "ProviderName", providerName)
-	return nil
-}
-
-func createStorageClusterPeer(ctx context.Context, client client.Client, logger *slog.Logger, mirrorPeer *multiclusterv1alpha1.MirrorPeer, clientInfoMap map[string]string) error {
-	logger = logger.With("MirrorPeer", mirrorPeer.Name)
-	items := mirrorPeer.Spec.Items
-	clientInfo := make([]odf.ClientInfo, 0)
-
-	for _, item := range items {
-		logger.Info("Fetching info for client", "ClientKey", utils.GetKey(item.ClusterName, item.StorageClusterRef.Name))
-		ci, err := odf.GetClientInfoFromConfigMap(clientInfoMap, utils.GetKey(item.ClusterName, item.StorageClusterRef.Name))
-		if err != nil {
-			return err
-		}
-		logger.Info("Client Info found", "ClientInfo", ci)
-		clientInfo = append(clientInfo, ci)
-	}
-
-	for i := range items {
-		var storageClusterPeerName string
-		var oppositeClient odf.ClientInfo
-		currentClient := clientInfo[i]
-		// Provider A StorageClusterPeer contains info of Provider B endpoint and ticket, hence this
-		if i == 0 {
-			oppositeClient = clientInfo[1]
-			storageClusterPeerName = getStorageClusterPeerName(oppositeClient.ProviderInfo.ProviderManagedClusterName)
-		} else {
-			oppositeClient = clientInfo[0]
-			storageClusterPeerName = getStorageClusterPeerName(oppositeClient.ProviderInfo.ProviderManagedClusterName)
-		}
-
-		// Provider B's onboarding token will be used for Provider A's StorageClusterPeer
-		logger.Info("Fetching onboarding ticket in with name and namespace", "Name", mirrorPeer.GetUID(), "Namespace", oppositeClient.ProviderInfo.ProviderManagedClusterName)
-		onboardingToken, err := fetchOnboardingTicket(ctx, client, oppositeClient, mirrorPeer)
-		if err != nil {
-			return fmt.Errorf("failed to fetch onboarding token for provider %s. %w", oppositeClient.ProviderInfo.ProviderManagedClusterName, err)
-		}
-
-		apiEndpoint := oppositeClient.ProviderInfo.StorageProviderPublicEndpoint
-		if apiEndpoint == "" {
-			logger.Error("'StorageProviderPublicEndpoint' not found. Using 'StorageProviderEndpoint' instead. It might not be accessible externally.")
-			apiEndpoint = oppositeClient.ProviderInfo.StorageProviderEndpoint
-		}
-
-		storageClusterPeer := ocsv1.StorageClusterPeer{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "StorageClusterPeer",
-				APIVersion: ocsv1.GroupVersion.String(),
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name: storageClusterPeerName,
-				// This provider A namespace on which the Storage object exists
-				Namespace: currentClient.ProviderInfo.NamespacedName.Namespace,
-			},
-			Spec: ocsv1.StorageClusterPeerSpec{
-				OnboardingToken: onboardingToken,
-				ApiEndpoint:     apiEndpoint,
-			},
-		}
-		storageClusterPeerJson, err := json.Marshal(storageClusterPeer)
-		if err != nil {
-			logger.Error("Failed to marshal StorageClusterPeer to JSON", "StorageClusterPeer", storageClusterPeerName)
-			return err
-		}
-
-		ownerRef := metav1.OwnerReference{
-			APIVersion: mirrorPeer.APIVersion,
-			Kind:       mirrorPeer.Kind,
-			Name:       mirrorPeer.Name,
-			UID:        mirrorPeer.UID,
-		}
-
-		// ManifestWork created for Provider A will be called storageclusterpeer-{ProviderA} since that is where Manifests will be applied
-		// Provider names are unique hence only 1 ManifestWork per ProviderCluster
-		manifestWorkName := fmt.Sprintf("storageclusterpeer-%s", currentClient.ProviderInfo.ProviderManagedClusterName)
-
-		// The namespace of Provider A is where this ManifestWork will be created on the hub
-		namespace := currentClient.ProviderInfo.ProviderManagedClusterName
-
-		manifesConfigOption := []workv1.ManifestConfigOption{
-			{
-				ResourceIdentifier: workv1.ResourceIdentifier{
-					Group:     ocsv1.GroupVersion.Group,
-					Resource:  "storageclusterpeers",
-					Name:      storageClusterPeer.Name,
-					Namespace: storageClusterPeer.Namespace,
-				},
-				FeedbackRules: []workv1.FeedbackRule{
-					{
-						Type: workv1.JSONPathsType,
-						JsonPaths: []workv1.JsonPath{
-							{
-								Name: "state",
-								Path: ".status.state",
-							},
-						},
-					},
-				},
-			},
-		}
-		operationResult, err := utils.CreateOrUpdateManifestWork(ctx, client, manifestWorkName, namespace, storageClusterPeerJson, manifesConfigOption, ownerRef)
-		if err != nil {
-			return err
-		}
-
-		logger.Info(fmt.Sprintf("ManifestWork was %s for StorageClusterPeer %s", operationResult, storageClusterPeerName))
-	}
-
-	return nil
-}
-
-func fetchOnboardingTicket(ctx context.Context, client client.Client, clientInfo odf.ClientInfo, mirrorPeer *multiclusterv1alpha1.MirrorPeer) (string, error) {
-	secretName := string(mirrorPeer.GetUID())
-	secretNamespace := clientInfo.ProviderInfo.ProviderManagedClusterName
-	tokenSecret := &corev1.Secret{}
-	if err := client.Get(ctx, types.NamespacedName{Name: secretName, Namespace: secretNamespace}, tokenSecret); err != nil {
-		if k8serrors.IsNotFound(err) {
-			return "", fmt.Errorf("secret %s not found in namespace %s", secretName, secretNamespace)
-		}
-		return "", fmt.Errorf("failed to fetch secret %s in namespace %s", secretName, secretNamespace)
-	}
-
-	tokenData, exists := tokenSecret.Data[utils.SecretDataKey]
-	if !exists {
-		return "", fmt.Errorf("token data not found in secret %s", secretName)
-	}
-	return string(tokenData), nil
-}
-
-func getStorageClusterPeerName(providerClusterName string) string {
-	// Provider A will have SCP named {ProviderB}-peer
-	return fmt.Sprintf("%s-peer", providerClusterName)
 }
 
 type ManagedClusterAddonConfig struct {
@@ -961,43 +724,6 @@ func (r *MirrorPeerReconciler) SetupWithManager(mgr ctrl.Manager) error {
 				utils.EventTypePredicate(true, true, true, true),
 			)).
 		Complete(r)
-}
-
-func GetNamespacedNameForClientS3Secret(pr multiclusterv1alpha1.PeerRef, mp *multiclusterv1alpha1.MirrorPeer, clientInfoMap map[string]string) (string, string, error) {
-	ci, err := odf.GetClientInfoFromConfigMap(clientInfoMap, utils.GetKey(pr.ClusterName, pr.StorageClusterRef.Name))
-	if err != nil {
-		return "", "", err
-	}
-	providerManagedClusterName := ci.ProviderInfo.ProviderManagedClusterName
-	pr1 := mp.Spec.Items[0]
-	pr2 := mp.Spec.Items[1]
-	s3SecretName := utils.CreateUniqueSecretNameForClient(providerManagedClusterName, utils.GetKey(pr1.ClusterName, pr1.StorageClusterRef.Name), utils.GetKey(pr2.ClusterName, pr2.StorageClusterRef.Name))
-	s3SecretNamespace := providerManagedClusterName
-
-	return s3SecretName, s3SecretNamespace, nil
-}
-
-func isProviderModePeeringDone(ctx context.Context, client client.Client, logger *slog.Logger, mirrorPeer *multiclusterv1alpha1.MirrorPeer, clientInfoMap map[string]string) (bool, error) {
-	isStorageClusterPeerManifestWorkCreated, err := checkStorageClusterPeerStatus(ctx, client, logger, mirrorPeer, clientInfoMap)
-	if err != nil {
-		logger.Error("failed to check if StorageClusterPeer have been created")
-		return false, err
-	}
-
-	logger.Info("StorageClusterPeer manifest work creation status", "isStorageClusterPeerManifestWorkCreated", isStorageClusterPeerManifestWorkCreated)
-	isClientPairingConfigMapCreated, err := checkClientPairingConfigMapStatus(ctx, client, logger, mirrorPeer, clientInfoMap)
-	if err != nil {
-		logger.Error("failed to check if client pair config map has been created")
-		return false, err
-	}
-
-	logger.Info("Client pairing ConfigMap creation status", "isClientPairingConfigMapCreated", isClientPairingConfigMapCreated)
-
-	allChecksPassed := isStorageClusterPeerManifestWorkCreated &&
-		isClientPairingConfigMapCreated
-
-	logger.Info("Provider mode peering status", "AllChecksPassed", allChecksPassed)
-	return allChecksPassed, nil
 }
 
 // isMirrorPeerValid validates the MirrorPeer CR
